tests and fmt
